Restructed a lot: api -> client, review utils -> pretalx utils
