+++ src/pytanis/__init__.py
@@ -2,9 +2,9 @@ from importlib.metadata import PackageNotFoundError, version
 
 import structlog.stdlib
 
-from .google import GoogleAPI
-from .helpdesk import HelpDeskAPI
-from .pretalx import PretalxAPI
+from .google import GSheetClient
+from .helpdesk import HelpDeskClient
+from .pretalx import PretalxClient
 
 try:
     __version__ = version("pytanis")
@@ -13,7 +13,7 @@ except PackageNotFoundError:  # pragma: no cover
 finally:
     del version, PackageNotFoundError
 
-__all__ = ["__version__", "GoogleAPI", "PretalxAPI", "HelpDeskAPI"]
+__all__ = ["__version__", "GSheetClient", "PretalxClient", "HelpDeskClient"]
 
 # transform structlog into a logging-friendly package
 # use `logging.basicConfig(level=logging.INFO, stream=sys.stdout)` as usual
